+++ src/main/java/com/prgrms/devcourse/springjpaboard/domain/user/api/UserController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.prgrms.devcourse.springjpaboard.domain.user.service.dto.UserSaveDto;
-import com.prgrms.devcourse.springjpaboard.domain.user.service.facade.UserFacade;
+import com.prgrms.devcourse.springjpaboard.domain.user.application.dto.UserSaveDto;
+import com.prgrms.devcourse.springjpaboard.domain.user.application.UserFacade;
 
 import lombok.RequiredArgsConstructor;
