+++ src/test/java/com/prgrms/devcourse/springjpaboard/domain/post/api/PostControllerDocTest.java
@@ -19,11 +19,11 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.prgrms.devcourse.springjpaboard.domain.post.Post;
 import com.prgrms.devcourse.springjpaboard.domain.post.repository.PostRepository;
-import com.prgrms.devcourse.springjpaboard.domain.post.service.dto.PostRequestDto;
-import com.prgrms.devcourse.springjpaboard.domain.post.service.dto.PostResponseDto;
-import com.prgrms.devcourse.springjpaboard.domain.post.service.dto.PostSaveDto;
-import com.prgrms.devcourse.springjpaboard.domain.post.service.dto.PostUpdateDto;
-import com.prgrms.devcourse.springjpaboard.domain.post.service.facade.PostFacade;
+import com.prgrms.devcourse.springjpaboard.domain.post.application.dto.PostRequestDto;
+import com.prgrms.devcourse.springjpaboard.domain.post.application.dto.PostResponseDto;
+import com.prgrms.devcourse.springjpaboard.domain.post.application.dto.PostSaveDto;
+import com.prgrms.devcourse.springjpaboard.domain.post.application.dto.PostUpdateDto;
+import com.prgrms.devcourse.springjpaboard.domain.post.application.PostFacade;
 import com.prgrms.devcourse.springjpaboard.domain.user.User;
 import com.prgrms.devcourse.springjpaboard.domain.user.repository.UserRepository;
