+++ src/test/java/com/prgrms/devcourse/springjpaboard/domain/post/application/PostServiceTest.java
@@ -1,4 +1,4 @@
-package com.prgrms.devcourse.springjpaboard.domain.post.service;
+package com.prgrms.devcourse.springjpaboard.domain.post.application;
 
 import java.util.List;
 import java.util.Optional;
